Fixed issues with content editor html previews having multiple panzooms
