+++ src/components/preview/page-preview-html.tsx
@@ -15,7 +15,7 @@ interface Props {
   layoutHtml: string;
   deviceModel: DeviceModel;
   resources: ExhibitionPageResource[];
-  selectedComponentId?: string;
+  borderedElementId?: string;
 }
 
 /**
@@ -25,19 +25,6 @@ interface PreviewProps {
   width: number;
   height: number;
 }
-
-/**
- * Preview Container styled component
- */
-const PreviewContainer = styled("div")(() => ({
-  display: "flex",
-  width: "100%",
-  minHeight: "100%",
-  justifyContent: "center",
-  alignContent: "center",
-  position: "relative",
-  backgroundColor: "#EFF0F1"
-}));
 
 /**
  * Preview styled component
@@ -61,9 +48,7 @@ const Preview = styled("iframe")(({ width, height }: PreviewProps) => ({
 /**
  * HTML Layouts Page Preview Component
  */
-const PagePreviewHtml = ({ deviceModel, screenOrientation, layoutHtml, resources, selectedComponentId }: Props) => {
-  const [showElementBorders, setShowElementBorders] = useState(false);
-
+const PagePreviewHtml = ({ deviceModel, screenOrientation, layoutHtml, resources, borderedElementId }: Props) => {
   if (!deviceModel) return null;
 
   const {
@@ -98,10 +83,10 @@ const PagePreviewHtml = ({ deviceModel, screenOrientation, layoutHtml, resources
    * @returns layout html with borders added
    */
   const addBorders = (html: string): string => {
-    if (selectedComponentId && showElementBorders) {
+    if (borderedElementId) {
       const htmlDocument = new DOMParser().parseFromString(html, "text/html");
 
-      const element = htmlDocument.getElementById(selectedComponentId);
+      const element = htmlDocument.getElementById(borderedElementId);
 
       if (element) {
         element.style["outline"] = "2px dashed #2196F3";
@@ -115,41 +100,10 @@ const PagePreviewHtml = ({ deviceModel, screenOrientation, layoutHtml, resources
   }
 
   return (
-    <PreviewContainer>
-      <PanZoom
-        minScale={0.1}
-        fitContent={true}
-        contentWidth={screenWidth}
-        contentHeight={screenHeight}
-        defaultPositionX={150}
-        defaultPositionY={150}
-      >
-        <Typography
-          sx={{
-            position: "absolute",
-            top: -20,
-            opacity: 0.6
-          }}
-        >
-          {deviceModel.model} / {screenHeight}x{screenWidth} /{" "}
-          {new Fraction((screenHeight ?? 0) / (screenWidth ?? 0)).toFraction().replace("/", ":")}
-        </Typography>
-        <Preview
-          srcDoc={ wrapHtmlLayout(addBorders(replaceResources(layoutHtml, resources))) }
-          {...getPreviewDimensions()}
-        />
-      </PanZoom>
-      <FormControlLabel
-        sx={{
-          position: "absolute",
-          bottom: 10
-        }}
-        label={strings.layoutEditorV2.preview.showElementBorders}
-        onChange={() => setShowElementBorders(!showElementBorders)}
-        value={showElementBorders}
-        control={<Switch checked={showElementBorders} color="secondary" />}
-      />
-    </PreviewContainer>
+    <Preview
+      srcDoc={wrapHtmlLayout(addBorders(replaceResources(layoutHtml, resources)))}
+      {...getPreviewDimensions()}
+    />
   );
 };
